+++ packages/riverpod_injected/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0+1] - 2026-09-10
+### Fixed
+- Added missing export
+
 ## [1.2.0] - 2026-09-08
 ### Added
 - Added `riverpodScope` method to create managed scopes for CLI based applications.
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Initial Release
 
+[1.2.0+1]: https://github.com/Skycoder42/riverpod_injected/compare/riverpod_injected-v1.2.0...riverpod_injected-v1.2.0+1
 [1.2.0]: https://github.com/Skycoder42/riverpod_injected/compare/riverpod_injected-v1.1.0...riverpod_injected-v1.2.0
 [1.1.0]: https://github.com/Skycoder42/riverpod_injected/compare/riverpod_injected-v1.0.0...riverpod_injected-v1.1.0
 [1.0.0]: https://github.com/Skycoder42/riverpod_injected/releases/tag/riverpod_injected-v1.0.0